+++ src/RuleSet/Php55.php
@@ -826,6 +826,11 @@ final class Php55 extends AbstractRuleSet implements ExplicitRuleSet
         return 'ergebnis (PHP 5.5)';
     }
 
+    public function rules(): array
+    {
+        return $this->rules;
+    }
+
     public function targetPhpVersion(): int
     {
         return 50500;