+++ src/RuleSet/Php71.php
@@ -829,6 +829,11 @@ final class Php71 extends AbstractRuleSet implements ExplicitRuleSet
         return 'ergebnis (PHP 7.1)';
     }
 
+    public function rules(): array
+    {
+        return $this->rules;
+    }
+
     public function targetPhpVersion(): int
     {
         return 70100;