+++ src/RuleSet/Php74.php
@@ -833,6 +833,11 @@ final class Php74 extends AbstractRuleSet implements ExplicitRuleSet
         return 'ergebnis (PHP 7.4)';
     }
 
+    public function rules(): array
+    {
+        return $this->rules;
+    }
+
     public function targetPhpVersion(): int
     {
         return 70400;