+++ CHANGELOG.md
@@ -17,6 +17,7 @@ For a full diff see [`5.15.1...main`][5.15.1...main].
 - Removed `$name` field from `Config\Ruleset\AbstractRuleSet` ([#865]), by [@localheinz]
 - Removed `$targetPhpVersion` field and `targetPhpVersion()` method from `Config\Ruleset\AbstractRuleSet` ([#866]), by [@localheinz]
 - Removed `name()` method from `Config\Ruleset\AbstractRuleSet` ([#867]), by [@localheinz]
+- Removed `rules()` method from `Config\Ruleset\AbstractRuleSet` ([#868]), by [@localheinz]
 
 ## [`5.15.1`][5.15.1]
 
@@ -1156,6 +1157,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#864]: https://github.com/ergebnis/php-cs-fixer-config/pull/864
 [#866]: https://github.com/ergebnis/php-cs-fixer-config/pull/866
 [#867]: https://github.com/ergebnis/php-cs-fixer-config/pull/867
+[#868]: https://github.com/ergebnis/php-cs-fixer-config/pull/868
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler